+++ tests/index.js
@@ -248,6 +248,161 @@ test('set: update should fail with bad data', async t => {
   }
 })
 
+test('set: content and profile idempotent with repeated values', async t => {
+  const p2p = createDb()
+  const sampleData = {
+    type: 'content',
+    title: 'demo',
+    description: 'lorem ipsum',
+    authors: [
+      'dat://3f70fe6b663b960a43a2c6c5a254c432196e2efa695e4b4e39779ae22e860e9d'
+    ],
+    parents: [
+      'dat://be53dcece25610c146b1617cf842593aa7ef134c6f771c2c145b9213deecf13a+4'
+    ]
+  }
+
+  const sampleProfile = {
+    type: 'profile',
+    title: 'professorX',
+    subtype: '',
+    main: 'test-profile.html',
+    avatar: './test.png',
+    follows: [
+      'dat://f7daadc2d624df738abbccc9955714d94cef656406f2a850bfc499c2080627d4'
+    ],
+    contents: [
+      'dat://00a4f2f18bb6cb4e9ba7c2c047c8560d34047457500e415d535de0526c6b4f23+12'
+    ]
+  }
+
+  const { rawJSON: content } = await p2p.init(sampleData)
+  const { rawJSON: profile } = await p2p.init(sampleProfile)
+
+  const ckey = content.url
+  const pkey = profile.url
+
+  const authors = sampleData.authors.concat(
+    'dat://3f70fe6b663b960a43a2c6c5a254c432196e2efa695e4b4e39779ae22e860e9d'
+  )
+
+  try {
+    // set content data parents field with some repeated values
+    await p2p.set({
+      url: ckey,
+      authors,
+      parents: sampleData.parents
+    })
+  } catch (err) {
+    t.ok(
+      err instanceof SDK.errors.ValidationError,
+      'trying to set duplicated values should throw ValidationError'
+    )
+  }
+
+  const { rawJSON: cUpdated } = await p2p.get(ckey)
+  t.same(cUpdated.authors, sampleData.authors, 'authors remains the same')
+  t.same(cUpdated.parents, sampleData.parents, 'parents remains the same')
+
+  // update profile with repeated values
+  const contents = [
+    'dat://3f70fe6b663b960a43a2c6c5a254c432196e2efa695e4b4e39779ae22e860alf',
+    'dat://00a4f2f18bb6cb4e9ba7c2c047c8560d34047457500e415d535de0526c6b4f23+12' // repeated value
+  ]
+  const follows = [
+    'dat://f7daadc2d624df738abbccc9955714d94cef656406f2a850bfc499c2080627d4', // repeated value
+    'dat://f7daadc2d624df738abbccc9955714d94cef656406f2a850bfc499c208062123'
+  ]
+
+  try {
+    await p2p.set({
+      url: pkey,
+      follows,
+      contents
+    })
+  } catch (err) {
+    t.same(
+      err.message,
+      'clone: Problems fetching external module',
+      'if module url is not found it will throw'
+    )
+  }
+
+  const { rawJSON: pUpdated } = await p2p.get(pkey)
+
+  t.same(pUpdated.follows, sampleProfile.follows, 'follows remains the same')
+  t.same(pUpdated.contents, sampleProfile.contents, 'contents remains the same')
+
+  await p2p.destroy()
+  t.end()
+})
+
+test('follows: must not self-reference', async t => {
+  const p2p = createDb()
+
+  t.plan(1)
+
+  const sampleProfile = {
+    type: 'profile',
+    title: 'professorX',
+    subtype: '',
+    main: 'test-profile.html',
+    avatar: './test.png',
+    follows: [
+      'dat://f7daadc2d624df738abbccc9955714d94cef656406f2a850bfc499c2080627d4'
+    ],
+    contents: [
+      'dat://00a4f2f18bb6cb4e9ba7c2c047c8560d34047457500e415d535de0526c6b4f23+12'
+    ]
+  }
+
+  const { rawJSON: profile } = await p2p.init(sampleProfile)
+
+  try {
+    await p2p.follow(profile.url, profile.url)
+  } catch (err) {
+    t.ok(
+      err instanceof SDK.errors.ValidationError,
+      'should throw when tries to self-reference'
+    )
+  }
+
+  await p2p.destroy()
+  t.end()
+})
+
+test('set: dont allow future parents versions nor self-reference', async t => {
+  const p2p = createDb()
+  const sampleData = {
+    type: 'content',
+    title: 'sample content',
+    description: 'lorem ipsum',
+    parents: [
+      'dat://be53dcece25610c146b1617cf842593aa7ef134c6f771c2c145b9213deecf13a+4'
+    ]
+  }
+  const { rawJSON } = await p2p.init(sampleData)
+  const ckey = rawJSON.url
+
+  try {
+    await p2p.set({
+      url: ckey,
+      parents: [
+        'dat://be53dcece25610c146b1617cf842593aa7ef134c6f771c2c145b9213deecf13a+4',
+        'dat://be53dcece25610c146b1617cf842593aa7ef134c6f771c2c145b9213deecf13a+4',
+        'dat://be53dcece25610c146b1617cf842593aa7ef134c6f771c2c145b9213deecf13a+4'
+      ]
+    })
+  } catch (err) {
+    t.ok(
+      err instanceof SDK.errors.ValidationError,
+      'invalid parents should throw ValidationError'
+    )
+  }
+  await p2p.destroy()
+  t.end()
+})
+
 test('update: check version change', async t => {
   const p2p = createDb()
   const sampleData = {
@@ -468,7 +623,7 @@ test('seed and publish', async t => {
   await p2p2.destroy(true, false)
 
   // call publish
-  await p2p.publish(contentKeyVersion, profile.url)
+  await p2p.publish(contentKeyVersionPrefix, profile.url)
 
   const { rawJSON } = await p2p.get(profile.url)
   t.same(
@@ -483,15 +638,20 @@ test('seed and publish', async t => {
   )
 
   // call publish again
-  await p2p.publish(contentKeyVersion, profile.url)
-  const { rawJSON: rawJSON2 } = await p2p.get(profile.url)
-  t.same(rawJSON.contents, rawJSON2.contents, 'publish is idempotent')
+  try {
+    await p2p.publish(contentKeyVersionPrefix, profile.url)
+  } catch (err) {
+    t.ok(
+      err instanceof SDK.errors.ValidationError,
+      'throws ValidationError with duplicated publish call'
+    )
+  }
 
   const dirs2 = await readdir(p2p.baseDir)
   t.same(
     dirs,
     dirs2,
-    'publish is idempotent (created directories remains the same)'
+    'repeated publish method call, created directories remains the same'
   )
   await p2p2.destroy(false, true)
   await p2p.destroy()
@@ -733,21 +893,49 @@ test('follow and unfollow a profile', async t => {
   }
 
   const { rawJSON: profileX } = await p2p.init(professorX)
-  const { rawJSON: profileY } = await p2p2.init(professorY)
+  const { rawJSON: profileY, metadata } = await p2p2.init(professorY)
 
   const followUrl = profileY.url
+  const followVersionedUrl = `${profileY.url}+${metadata.version}`
+
   t.equal(profileX.follows.length, 0, 'Initially follows should be empty')
-  // call follow
+
+  // call follow (unversioned)
   await p2p.follow(profileX.url, followUrl)
 
+  // follow an nonexistant profile should throw
+  try {
+    await p2p.follow(
+      profileX.url,
+      'dat://be53dcece25610c146b1617cf842593aa7ef134c6f771c2c145b9213deecf13a'
+    )
+  } catch (err) {
+    t.same(
+      err.message,
+      'clone: Problems fetching external module',
+      'Trying to follow a nonexistant module should throw'
+    )
+  }
+
   const { rawJSON: profileXUpdated } = await p2p.get(profileX.url)
-  t.equal(profileXUpdated.follows.length, 1)
+  t.equal(profileXUpdated.follows.length, 1, 'following 1 new profile')
   t.same(
     profileXUpdated.follows,
     [followUrl],
     'follows property should contain target profile url'
   )
 
+  // unfollow versioned profile url which does not exists
+  await p2p.unfollow(profileX.url, followVersionedUrl)
+
+  const { rawJSON: profileXSame } = await p2p.get(profileX.url)
+  t.equal(
+    profileXSame.follows.length,
+    1,
+    'unfollow does nothing when profile does not match (versioned)'
+  )
+
+  // unfollow unversioned profile url
   await p2p.unfollow(profileX.url, followUrl)
 
   const { rawJSON: profileXFinal } = await p2p.get(profileX.url)
@@ -755,6 +943,37 @@ test('follow and unfollow a profile', async t => {
     profileXFinal.follows.length,
     0,
     'unfollow removes the target profile'
+  )
+
+  // call follow (versioned)
+  await p2p.follow(profileX.url, followVersionedUrl)
+
+  const { rawJSON: profileXUpdated2 } = await p2p.get(profileX.url)
+  t.equal(profileXUpdated2.follows.length, 1, 'following 1 new profile')
+  t.same(
+    profileXUpdated2.follows,
+    [followVersionedUrl],
+    'follows property should contain target profile versioned url'
+  )
+
+  // unfollow unversioned profile but only versioned is being followed
+  await p2p.unfollow(profileX.url, followUrl)
+
+  const { rawJSON: profileXNoChange } = await p2p.get(profileX.url)
+  t.equal(
+    profileXNoChange.follows.length,
+    1,
+    'unfollow does nothing when the target profile does not match (unversioned)'
+  )
+
+  // unfollow versioned profile url
+  await p2p.unfollow(profileX.url, followVersionedUrl)
+
+  const { rawJSON: profileXFinal2 } = await p2p.get(profileX.url)
+  t.equal(
+    profileXFinal2.follows.length,
+    0,
+    'unfollow removes the target versioned profile '
   )
 
   await p2p.destroy()