+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - verify method improvements #97
+- update to module specs 0.2.3 #99
 
 ### Changed
 - make parents dat url validation strict #96
+- refactor set method so follow and publish calls pass through it
 - switch to GitHub actions for CI #114
 
 ### Fixed